use push instead of merge
There seemed to be a problem with merging master into core-api branch.  I'm not sure if the manual changes to core-api affected this.

However I realized a 'merge' is conceptually bringing together two sources of truth, but that's not really what we're doing here.  We really just want to take main as our single source of truth and copy it completely into core-api and channel-api branches.  I think this force push from main to the other branches is simpler and actually matches our mental model a little better.